refactor(gfn): extract pure webrtc helpers into focused modules
Move diagnostics types, stream lag classification, mouse input utils,
controller overlay gate, and stats helpers out of webrtcClient.ts while
re-exporting the same public symbols for existing import paths.
